Widen polygon scale to 16 unsigned bits
Even though the DrawSpritePolygon instruction only reads a single byte
for a constant scale, [the reference implementation][1] internally
handles it as a 16-bit integer and will respect the whole 16-bit value
from a register (treating it as an unsigned).

To scale polygon coordinates, Another World multiplies them by the
scale value and then divides the result by 64: so e.g. a scale of 32
means 0.5x, a scale of 96 means 1.5x, and a scale of 255 means ~4x.
This approach allowed the game to do fractional scaling without
floating-point arithmetic.

Since scale values in code must therefore be premultiplied by 64,
it's reasonable to think that scales in registers would sometimes
exceed 255.

[1]: https://github.com/fabiensanglard/Another-World-Bytecode-Interpreter/blob/dea6914a82f493cb329188bcffa46b9d0b234ea6/src/video.h#L91

+++ vm/video.zig
@@ -23,7 +23,7 @@ pub const PolygonAddress = u16;
 
 /// The scale at which to render a polygon.
 /// TODO: document the observed ranges and default value for this.
-pub const PolygonScale = u8;
+pub const PolygonScale = u16;
 
 const log_unimplemented = @import("../utils/logging.zig").log_unimplemented;
